backend/authentication-service/requirements.txt
blinker==1.7.0
click==8.1.7
colorama==0.4.6
flask==3.0.3
Flask-Cors==4.0.0
Flask-JWT-Extended==4.6.0
flask-sqlalchemy==3.1.1
greenlet==3.0.3
importlib-metadata==7.1.0
itsdangerous==2.2.0
Jinja2==3.1.3
MarkupSafe==2.1.5
psycopg2==2.9.9
PyJWT==2.8.0
pyotp==2.9.0
python-dotenv==1.0.1
SQLAlchemy==2.0.29
typing-extensions==4.11.0
werkzeug==3.0.2
zipp==3.18.1
requests==2.31.0
